requirements.txt
fastapi==0.111.0
httpx==0.27.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
anthropic==0.31.1
python-dotenv==1.0.0
pytest==8.2.2
pluggy==1.5.0
Faker==26.0.0
pytest-asyncio==0.23.7
